prompts/requirements.txt
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
python-dotenv
colorama
